+++ tools/Azure.Mcp.Tools.AzureManagedLustre/src/AzureManagedLustreSetup.cs
@@ -31,5 +31,6 @@ public class AzureManagedLustreSetup : IAreaSetup
         fileSystem.AddCommand("list", new FileSystemListCommand(loggerFactory.CreateLogger<FileSystemListCommand>()));
         fileSystem.AddCommand("required-subnet-size", new FileSystemSubnetSizeCommand(loggerFactory.CreateLogger<FileSystemSubnetSizeCommand>()));
         fileSystem.AddCommand("create", new FileSystemCreateCommand(loggerFactory.CreateLogger<FileSystemCreateCommand>()));
+        fileSystem.AddCommand("update", new FileSystemUpdateCommand(loggerFactory.CreateLogger<FileSystemUpdateCommand>()));
     }
 }