+++ servers/Azure.Mcp.Server/CHANGELOG.md
@@ -8,6 +8,7 @@ The Azure MCP Server updates automatically by default whenever a new release com
 
 - Added the following Azure Managed Lustre commands: [[#50](https://github.com/microsoft/mcp/issues/50)]
   - `azmcp_azuremanagedlustre_filesystem_create`: Create an Azure Managed Lustre filesystems.
+  - `azmcp_azuremanagedlustre_filesystem_update`: Update an Azure Managed Lustre filesystems.
 
 ### Breaking Changes